move sendAccountRemovedToSegment.ts into analytics.ts
